refactor: move customer models to a specific package
