+++ src/main/java/com/abacatepay/model/customer/Customer.java
@@ -1,0 +1,11 @@
+package com.abacatepay.model.customer;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Customer {
+    private String id;
+    private CustomerMetadata metadata;
+}
