+++ src/main/java/com/abacatepay/model/customer/CustomerMetadata.java
@@ -1,4 +1,4 @@
-package com.abacatepay.model.billing;
+package com.abacatepay.model.customer;
 
 import lombok.Data;
